+++ actix-web-lab/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Re-work `Json` deserialization error handling.
+- Re-work `Json` extractor error handling.
+- Re-work `UrlEncodedForm` extractor error handling.
 - Upgrade to edition 2024.
 - Minimum supported Rust version (MSRV) is now 1.85.